modularize routes, update app permissions for gvm ease of use, add config file
